unsealed file size ecall and ocall

+++ secureUI/sgxware.h
@@ -67,7 +67,7 @@ public:
     bool         refreshEPG(void);
     DRM_status_t readMovieChunk(unsigned int movie_id, size_t movie_offset, size_t chunk_size, unsigned char* dest);
     bool         writeAppLog (unsigned char* data, size_t length);
-    bool        getFileName(unsigned int id, size_t movie_name_size, char *movie_name);
+    bool        getFileName(size_t id, size_t movie_name_size, char *movie_name);
 	bool        inplaceDecrypt(qint64 size, void* data);
 	bool        getFileSize(size_t movie_id, size_t* fsize);
